medmnist_web/requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
python-multipart
torchxrayvision
numpy
